+++ ivy/functional/backends/torch/extensions/layers.py
@@ -1,0 +1,159 @@
+# global
+from typing import Optional, Union, Tuple, Literal, List
+from numbers import Number
+import torch
+import math
+
+# local
+import ivy
+
+
+def flatten(
+    x: torch.Tensor,
+    /,
+    *,
+    start_dim: Optional[int] = 0,
+    end_dim: Optional[int] = -1,
+    out: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    return torch.flatten(x, start_dim=start_dim, end_dim=end_dim)
+
+
+def vorbis_window(
+    window_length: torch.tensor,
+    *,
+    dtype: Optional[torch.dtype] = torch.float32,
+    out: Optional[torch.tensor] = None,
+) -> torch.tensor:
+    return torch.tensor(
+        [
+            round(
+                math.sin(
+                    (ivy.pi / 2) * (math.sin(ivy.pi * (i) / (window_length * 2)) ** 2)
+                ),
+                8,
+            )
+            for i in range(1, window_length * 2)[0::2]
+        ],
+        dtype=dtype,
+    )
+
+
+vorbis_window.support_native_out = False
+
+
+def hann_window(
+    window_length: int,
+    periodic: Optional[bool] = True,
+    dtype: Optional[torch.dtype] = None,
+    *,
+    out: Optional[torch.tensor] = None,
+) -> torch.tensor:
+    return torch.hann_window(
+        window_length,
+        periodic=periodic,
+        dtype=dtype,
+        layout=torch.strided,
+        device=None,
+        requires_grad=None,
+    )
+
+
+hann_window.support_native_out = False
+
+
+# noinspection PyUnresolvedReferences
+def max_pool2d(
+    x: torch.Tensor,
+    kernel: Union[int, Tuple[int], Tuple[int, int]],
+    strides: Union[int, Tuple[int], Tuple[int, int]],
+    padding: str,
+    /,
+    *,
+    data_format: str = "NHWC",
+    out: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    if isinstance(strides, int):
+        strides = (strides, strides)
+    elif len(strides) == 1:
+        strides = (strides[0], strides[0])
+
+    if data_format == "NHWC":
+        x = x.permute(0, 3, 1, 2)
+    x_shape = list(x.shape[2:])
+    pad_h = ivy.handle_padding(x_shape[0], strides[0], kernel[0], padding)
+    pad_w = ivy.handle_padding(x_shape[1], strides[1], kernel[1], padding)
+    x = torch.nn.functional.pad(
+        x,
+        [pad_w // 2, pad_w - pad_w // 2, pad_h // 2, pad_h - pad_h // 2],
+        value=float("-inf"),
+    )
+    if padding != "VALID" and padding != "SAME":
+        raise ivy.exceptions.IvyException(
+            "Invalid padding arg {}\n"
+            'Must be one of: "VALID" or "SAME"'.format(padding)
+        )
+    res = torch.nn.functional.max_pool2d(x, kernel, strides, 0)
+    if data_format == "NHWC":
+        return res.permute(0, 2, 3, 1)
+    return res
+
+
+max_pool2d.unsupported_dtypes = ("bfloat16", "float16")
+
+
+def pad(
+    x: torch.Tensor,
+    /,
+    pad_width: Tuple[int],
+    *,
+    mode: Optional[Literal["constant", "reflect", "edge", "wrap"]] = "constant",
+    stat_length: Optional[Union[torch.Tensor, int]] = None,
+    constant_values: Optional[Number] = 0,
+    end_values: Optional[Number] = 0,
+    reflect_type: Optional[Literal["even", "odd"]] = "even",
+    out: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    if x.shape == ():
+        x = x.unsqueeze(0)
+    if isinstance(pad_width, torch.Tensor):
+        pad_width = pad_width.detach().cpu().numpy().tolist()
+    pad_width.reverse()
+    pad_width_flat: List[int] = list()
+    for pad_width_sec in pad_width:
+        for item in pad_width_sec:
+            pad_width_flat.append(item)
+    if mode == "constant":
+        return torch.nn.functional.pad(
+            x,
+            pad_width_flat,
+            mode=mode,
+            value=constant_values,
+        )
+    else:
+        x = x.unsqueeze(dim=0)
+        if mode == "edge":
+            mode = "replicate"
+        elif mode == "wrap":
+            mode = "circular"
+            x = x.unsqueeze(dim=0)
+        return torch.nn.functional.pad(x, pad_width_flat, mode=mode).squeeze()
+
+
+def kaiser_window(
+    window_length: int,
+    periodic: bool = True,
+    beta: float = 12.0,
+    *,
+    dtype: Optional[torch.dtype] = None,
+    out: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    return torch.kaiser_window(
+        window_length,
+        periodic,
+        beta,
+        dtype=dtype,
+        layout=torch.strided,
+        device=None,
+        requires_grad=False,
+    )
