+++ ivy/functional/backends/torch/extensions/manipulation.py
@@ -1,0 +1,45 @@
+from typing import Optional, Union, Sequence
+import torch
+
+
+def moveaxis(
+    a: torch.Tensor,
+    source: Union[int, Sequence[int]],
+    destination: Union[int, Sequence[int]],
+    /,
+    *,
+    out: Optional[torch.Tensor] = None,
+) -> torch.Tensor:
+    return torch.moveaxis(a, source, destination)
+
+
+moveaxis.support_native_out = False
+
+
+def heaviside(
+    x1: torch.tensor,
+    x2: torch.tensor,
+    /,
+    *,
+    out: Optional[torch.tensor] = None,
+) -> torch.tensor:
+    return torch.heaviside(
+        x1,
+        x2,
+        out=out,
+    )
+
+
+heaviside.support_native_out = True
+
+
+def flipud(
+    m: torch.Tensor,
+    /,
+    *,
+    out: Optional[torch.tensor] = None,
+) -> torch.tensor:
+    return torch.flipud(m)
+
+
+flipud.support_native_out = False
